arm64: add support for appended dtb
this is achieved by passing compressed kernel to kexec load, which will
then scan it for dtb and use it instead of using dtb from original
boot.img. In case this doesn't work, it will fallback to the old method.
Some function signatures have been changed to allow passing compressed
kernel buf in addition to uncompressed one.
The reason we need that is, the decompression of zImage causes us to
lose the appended dtb structures, which are classified as
trailing garbage for zlib.
Code for reading appended dtb was in large taken from this commit
https://github.com/Tasssadar/kexec-tools/commit/b5fa96dff3d8cc4ec5d5ed335d0c6acddfc54af5